[TASK] Add AddionalConfiguration content at installation
